+++ engine/src/main/java/net/jqwik/engine/hooks/statistics/StatisticsHook.java
@@ -2,6 +2,7 @@ package net.jqwik.engine.hooks.statistics;
 
 import java.util.*;
 import java.util.function.*;
+import java.util.logging.*;
 import java.util.stream.*;
 
 import net.jqwik.api.*;
@@ -13,6 +14,8 @@ import net.jqwik.engine.hooks.*;
 import net.jqwik.engine.support.*;
 
 public class StatisticsHook implements AroundPropertyHook {
+
+	private static final Logger LOG = Logger.getLogger(StatisticsHook.class.getName());
 
 	private static final Supplier<Map<String, StatisticsCollectorImpl>> STATISTICS_MAP_SUPPLIER =
 		() -> new LinkedHashMap<String, StatisticsCollectorImpl>() {
@@ -34,7 +37,8 @@ public class StatisticsHook implements AroundPropertyHook {
 		PropertyExecutionResult testExecutionResult = property.execute();
 
 		Map<String, StatisticsCollectorImpl> collectors = collectorsStore.get();
-		createStatisticsReports(collectors, context);
+		boolean isFailure = testExecutionResult.status() == PropertyExecutionResult.Status.FAILED;
+		createStatisticsReports(collectors, context, isFailure);
 		if (testExecutionResult.status() == PropertyExecutionResult.Status.SUCCESSFUL) {
 			return checkCoverages(testExecutionResult, collectors.values());
 		}
@@ -56,7 +60,11 @@ public class StatisticsHook implements AroundPropertyHook {
 		}
 	}
 
-	private void createStatisticsReports(Map<String, StatisticsCollectorImpl> collectors, PropertyLifecycleContext context) {
+	private void createStatisticsReports(
+		Map<String, StatisticsCollectorImpl> collectors,
+		PropertyLifecycleContext context,
+		boolean isFailure
+	) {
 		List<StatisticsReport> statisticsReportAnnotations = JqwikAnnotationSupport.findRepeatableAnnotationOnElementOrContainer(
 			context.targetMethod(),
 			StatisticsReport.class
@@ -66,7 +74,7 @@ public class StatisticsHook implements AroundPropertyHook {
 			collectors.entrySet().stream()
 					  .map(entry -> {
 						  String label = entry.getKey();
-						  return Tuple.of(label, entry.getValue(), determineFormat(label, statisticsReportAnnotations, context));
+						  return Tuple.of(label, entry.getValue(), determineFormat(label, statisticsReportAnnotations, context, isFailure));
 					  })
 					  .collect(Collectors.toSet());
 
@@ -76,27 +84,38 @@ public class StatisticsHook implements AroundPropertyHook {
 	private StatisticsReportFormat determineFormat(
 		String label,
 		List<StatisticsReport> statisticsReportAnnotations,
-		PropertyLifecycleContext context
+		PropertyLifecycleContext context,
+		boolean isFailure
 	) {
 		boolean defaultReportFormatSet = false;
 		StatisticsReportFormat defaultReportFormat = new StandardStatisticsReportFormat();
 
 		for (StatisticsReport annotation : statisticsReportAnnotations) {
 			if (annotation.label().equals(label)) {
-				return createReportFormat(annotation, context);
+				return createReportFormat(annotation, context, isFailure);
 			}
 			if (!defaultReportFormatSet && annotation.label().equals(StatisticsReport.ALL_LABELS)) {
-				defaultReportFormat = createReportFormat(annotation, context);
+				defaultReportFormat = createReportFormat(annotation, context, isFailure);
 				defaultReportFormatSet = true;
 			}
 		}
 		return defaultReportFormat;
 	}
 
-	private StatisticsReportFormat createReportFormat(StatisticsReport annotation, PropertyLifecycleContext context) {
+	private StatisticsReportFormat createReportFormat(StatisticsReport annotation, PropertyLifecycleContext context, boolean isFailure) {
 		if (annotation.value() == StatisticsReportMode.OFF) {
+			if (annotation.onFailureOnly()) {
+				String message = String.format(
+					"@StatisticsReport(onFailureOnly = true, value = StatisticsReport.StatisticsReportMode.OFF) " +
+						"does not make sense.%n" +
+						"You should change 'onFailureOnly' or 'value'."
+				);
+				LOG.warning(message);
+			}
 			return null;
-		} else if (annotation.value() == StatisticsReportMode.PLUG_IN) {
+		} else if (annotation.onFailureOnly() && !isFailure) {
+			return null;
+		} else if (annotation.value() == StatisticsReportMode.PLUG_IN && !Objects.equals(annotation.format(), None.class)) {
 			return JqwikReflectionSupport.newInstanceInTestContext(annotation.format(), context.testInstance());
 		} else {
 			return new StandardStatisticsReportFormat();