+++ engine/src/test/java/net/jqwik/api/ArrayArbitraryTests.java
@@ -30,7 +30,7 @@ class ArrayArbitraryTests {
 	}
 
 	@Example
-	@StatisticsReport(StatisticsReport.StatisticsReportMode.OFF)
+	@StatisticsReport(onFailureOnly = true)
 	void withSizeDistribution(@ForAll Random random) {
 		Arbitrary<Integer> integerArbitrary = Arbitraries.integers();
 		ArrayArbitrary<Integer, Integer[]> arbitrary =