+++ engine/src/test/java/net/jqwik/api/SetArbitraryTests.java
@@ -109,7 +109,7 @@ class SetArbitraryTests {
 	}
 
 	@Example
-	@StatisticsReport(StatisticsReport.StatisticsReportMode.OFF)
+	@StatisticsReport(onFailureOnly = true)
 	void withSizeDistribution(@ForAll Random random) {
 		Arbitrary<Integer> integerArbitrary = Arbitraries.integers();
 		SetArbitrary<Integer> arbitrary =