+++ engine/src/test/java/net/jqwik/engine/properties/arbitraries/DefaultStringArbitraryTests.java
@@ -12,6 +12,7 @@ import net.jqwik.api.statistics.*;
 
 import static net.jqwik.testing.TestingSupport.*;
 
+@StatisticsReport(onFailureOnly = true)
 class DefaultStringArbitraryTests implements GenericEdgeCasesProperties {
 
 	@Override
@@ -164,7 +165,6 @@ class DefaultStringArbitraryTests implements GenericEdgeCasesProperties {
 	}
 
 	@Example
-	@StatisticsReport(StatisticsReport.StatisticsReportMode.OFF)
 	void withLengthDistribution(@ForAll Random random) {
 		StringArbitrary arbitrary = this.arbitrary.ofMaxLength(100)
 												  .withLengthDistribution(RandomDistribution.uniform());
@@ -254,7 +254,6 @@ class DefaultStringArbitraryTests implements GenericEdgeCasesProperties {
 	@Group
 	class Coverage {
 		@Property
-		@StatisticsReport(StatisticsReport.StatisticsReportMode.OFF)
 		void randomStringsShouldContainZeroChar(@ForAll @StringLength(min = 1, max = 20) String aString) {
 			Statistics.label("contains 0")
 					  .collect(aString.contains("\u0000"))
@@ -265,7 +264,6 @@ class DefaultStringArbitraryTests implements GenericEdgeCasesProperties {
 		}
 
 		@Property(edgeCases = EdgeCasesMode.NONE)
-		@StatisticsReport(StatisticsReport.StatisticsReportMode.OFF)
 		void evenLongStringsShouldSometimesGenerateNoDuplicates(@ForAll @StringLength(500) String aString) {
 			Statistics.label("duplicates")
 					  .collect(hasDuplicate(aString))
@@ -293,7 +291,6 @@ class DefaultStringArbitraryTests implements GenericEdgeCasesProperties {
 		}
 
 		@Property(generation = GenerationMode.RANDOMIZED, edgeCases = EdgeCasesMode.NONE)
-		@StatisticsReport(StatisticsReport.StatisticsReportMode.OFF)
 		void stringFromSeveralCharacterGroups_HaveRandomDistributionBySize(@ForAll("a to z, 1 to 3") String aString) {
 			char onlyChar = aString.charAt(0);
 
@@ -317,7 +314,6 @@ class DefaultStringArbitraryTests implements GenericEdgeCasesProperties {
 		}
 
 		@Property(tries = 2000, edgeCases = EdgeCasesMode.NONE)
-		@StatisticsReport(StatisticsReport.StatisticsReportMode.OFF)
 		void randomStringsWithRepeatedCharsShouldGenerateDuplicatesAndRepetitions(
 			@ForAll("withRepeatedChars") @StringLength(min = 10, max = 20) String aString
 		) {